fix(garden): moving delete button
moving the Garden delete button off of the Garden page and in to the edit modal

fix #39
